refactor: 북마크 API 리팩토링 (#114)
